Use platform-specific read_at to read the files.
All frfs::File uses the same source underneath. The sources are created
with try_clone. Therefore when one file is readed, all other files'
cursor will be changed. This is not a desirable behavior.

While there are no cross-platform way to avoid this besides reopening
the source, we can use read_at/seek_read methods in the FileExt. It is a
platform-specific extension, but avaliable on Windows, Unix, and WASI.

The implementation on Unix of read_at is pread/pread64, and the one on
Windows is NtReadFile. They are all reliable methods and will not read
or write the file cursor.

Using read_at, we need to manage the own offset in the frfs::File.

This implementation is not suitable for try_clone. Therefore we removed
this method.